Fix requested changes
